backend/requirements.txt
fastapi
uvicorn
pymongo
motor
pyjwt
passlib
python-dotenv
redis
python-multipart